Created a static solution for double navigation bars while extending from base
Created a static solution for double navigation bars while extending from base, basically involves creating another method where the value is already frozen as a literal and not variable. This was all to allow a single page view of everything on the site if somebody just wanted to scroll through the home real quick.