Update styles css, project description for SPAN
One pitfall is the newlines and formatting for project description but ease of access is worth it, I could always just go back and format it after the freeze if I wish- or- simply pull the .md from git if it has an applicable git asset link. I may add this code in another release.