Blog is up!
Blog is up! Files are generated for both live and frozen versions for pages. I wanted a blog and got one! Later on I will add a search feature and etc, comments and all that (if possible through pages) but I wanted a spot to share my ideas in a nice easy format. Here we are!